fin_partie1

+++ OCTA_Projet_Gestion_Commerciale.Data/Infrastructure/IRepository.cs
@@ -12,7 +12,7 @@ namespace OCTA_Projet_Gestion_Commerciale.Data.Infrastructure
         // Marks an entity as new
         void Add(T entity);
         // Marks an entity as modified
-         void Update(T entity);
+        void Update(T entity);
         void Update(object idSource, T entity);
         // Marks an entity to be removed
         void Delete(object idSource, T entity);
@@ -20,12 +20,36 @@ namespace OCTA_Projet_Gestion_Commerciale.Data.Infrastructure
         //void Delete(T entity);
         void Delete(Expression<Func<T, bool>> where);
         // Get an entity by int id
-        T GetById(long id);
+        T GetById(int id);
         // Get an entity using delegate
         T Get(Expression<Func<T, bool>> where);
         // Gets all entities of type T
         IEnumerable<T> GetAll();
         // Gets entities using delegate
         IEnumerable<T> GetMany(Expression<Func<T, bool>> where);
+
+
+
+
+
+
+        //// Marks an entity as new
+        //void Add(T entity);
+        //// Marks an entity as modified
+        // void Update(T entity);
+        //void Update(object idSource, T entity);
+        //// Marks an entity to be removed
+        //void Delete(object idSource, T entity);
+
+        ////void Delete(T entity);
+        //void Delete(Expression<Func<T, bool>> where);
+        //// Get an entity by int id
+        //T GetById(long id);
+        //// Get an entity using delegate
+        //T Get(Expression<Func<T, bool>> where);
+        //// Gets all entities of type T
+        //IEnumerable<T> GetAll();
+        //// Gets entities using delegate
+        //IEnumerable<T> GetMany(Expression<Func<T, bool>> where);
     }
 }
